Switched to fixed-width integer types

+++ Types.h
@@ -1,12 +1,13 @@
 #pragma once
+#include <cstdint>
 
 typedef signed char         s8;
-typedef signed short        s16;
-typedef signed long         s32;
-typedef signed long long    s64;
+typedef int16_t             s16;
+typedef int32_t             s32;
+typedef int64_t             s64;
 typedef unsigned char       u8;
-typedef unsigned short      u16;
-typedef unsigned long       u32;
-typedef unsigned long long  u64;
+typedef uint16_t            u16;
+typedef uint32_t            u32;
+typedef uint64_t            u64;
 typedef float               f32;
 typedef double              f64;
